pyproject.toml
[tool.poetry]
name = "xmppbot"
version = "2.0.0"
description = "A framework for writing Jabber/XMPP bots"
authors = ["s-nt-s <santos82h@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slixmpp = "^1.8.3"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
autopep8 = "^2.0.2"
autoflake = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
